Split StakePortfolioCard
Split StakePortfolioCard component to two:
- NewStakeCard
- StakePortfolioCard

We basically extract the form for the new stake creation to a new component.
