+++ src/pages/Staking/StakedPortfolioCard.tsx
@@ -1,5 +1,5 @@
-import { ComponentProps, FC, useMemo } from "react"
-import { HStack, Divider } from "@chakra-ui/react"
+import { ComponentProps, FC } from "react"
+import { HStack } from "@chakra-ui/react"
 import {
   BodyLg,
   BodyMd,
@@ -9,37 +9,15 @@ import {
 } from "@threshold-network/components"
 import InfoBox from "../../components/InfoBox"
 import TokenBalance from "../../components/TokenBalance"
-import { StakingContractLearnMore } from "../../components/ExternalLink"
-import { TokenAmountForm } from "../../components/Forms"
 import { useStakingState } from "../../hooks/useStakingState"
-import { useModal } from "../../hooks/useModal"
 import { useTokenBalance } from "../../hooks/useTokenBalance"
-import { ModalType, Token } from "../../enums"
+import { Token } from "../../enums"
 import { formatTokenAmount } from "../../utils/formatAmount"
-import { useMinStakeAmount } from "../../hooks/useMinStakeAmount"
 
 const StakedPortfolioCard: FC<ComponentProps<typeof Card>> = (props) => {
-  const { openModal } = useModal()
   const tBalance = useTokenBalance(Token.T)
-  const { minStakeAmount, isLoading, hasError } = useMinStakeAmount()
-
-  const openStakingModal = async (stakeAmount: string) => {
-    openModal(ModalType.StakingChecklist, { stakeAmount })
-  }
 
   const { stakedBalance } = useStakingState()
-
-  const placeholder = useMemo(() => {
-    if (hasError) {
-      return "Error fetching min stake"
-    }
-
-    return `Minimum stake ${
-      isLoading || minStakeAmount === undefined
-        ? "loading..."
-        : `${formatTokenAmount(minStakeAmount)} T`
-    }`
-  }, [isLoading, hasError, minStakeAmount])
 
   return (
     <Card h="fit-content" {...props}>
@@ -56,19 +34,9 @@ const StakedPortfolioCard: FC<ComponentProps<typeof Card>> = (props) => {
         />
       </InfoBox>
       <HStack mt="8" justify="space-between" w="100%">
-        <BoxLabel>Wallet</BoxLabel>
+        <BoxLabel status="secondary">Wallet</BoxLabel>
         <BodyLg>{formatTokenAmount(tBalance)} T</BodyLg>
       </HStack>
-      <Divider my="6" />
-      <TokenAmountForm
-        onSubmitForm={openStakingModal}
-        label="Stake Amount"
-        submitButtonText="New Stake"
-        maxTokenAmount={tBalance}
-        placeholder={placeholder}
-        minTokenAmount={minStakeAmount}
-      />
-      <StakingContractLearnMore mt="3" />
     </Card>
   )
 }
